Fixes unarchive call
The role worked because of the fact that workstation and provisioned
host were the same.
unarchive requires an archive file on the local workstations' disk
to be transfered to the host before extracting the archive.

Download the archive to the local system using local_action: get_url
further allows to download an archive once and deploy it on many
machines.